key_gen.py
from ecdsa import SigningKey, SECP256k1
import sha3

def new_address():
    keccak = sha3.keccak_256()

    priv = SigningKey.generate(curve=SECP256k1)
    pub = priv.get_verifying_key().to_string()

    keccak.update(pub)
    address = keccak.hexdigest()[24:]

    return '0x' + address
