requirements.txt
flask
flask_restful
flask_cors
paho-mqtt
pysha3
ecdsa

